rag_backend/requirements-production.txt
# RAG Production - Pre-computed Index
# ====================================
# Uses pre-computed FAISS index + OpenAI embeddings for queries
# Memory: ~100MB (vs 1GB+ with Sentence Transformers)

# FastAPI Server
fastapi>=0.100.0
uvicorn[standard]>=0.25.0
pydantic>=2.0.0

# FAISS for vector search (CPU only, no heavy ML deps)
faiss-cpu>=1.7.4

# OpenAI for query embeddings + chatbot
openai>=1.0.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.24.0
orjson>=3.9.0
numpy>=1.24.0

# CORS & Streaming
python-multipart>=0.0.6
sse-starlette>=1.0.0
